playground/granite_lab/requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
pypdf>=3.0.0
pillow>=10.0.0
requests>=2.31.0
